Removed hack, compareDocumentPosition implmented from zero, base by https://github.com/hunterhacker/jdom Added parentNode to all nodes except Document Impprovved DOMTokenList Clean Unit test

+++ LoboUnitTest/src/test/java/org/loboevolution/domts/level3/nodecomparedocumentposition22Test.java
@@ -64,7 +64,7 @@ public class nodecomparedocumentposition22Test extends LoboUnitTest {
         notation = (Notation) notationsMap.getNamedItem("notation1");
         entityPosition = entity.compareDocumentPosition(notation);
         assertEquals(4, entityPosition);
-        notationPosition = (int) notation.compareDocumentPosition(entity);
+        notationPosition = notation.compareDocumentPosition(entity);
         assertEquals(2, notationPosition);
     }
 }
